requirements.txt
# ==========================================
# ☕ Coffee AX Master Hub - Requirements
# ==========================================
# Streamlit 코어
streamlit>=1.28.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 시각화
plotly>=5.18.0
matplotlib>=3.7.0
wordcloud>=1.9.0

# 지도
folium>=0.14.0
streamlit-folium>=0.15.0

# 금융 데이터
yfinance>=0.2.28

# API 통신
requests>=2.31.0
urllib3>=2.0.0

# 환경 변수
python-dotenv>=1.0.0

# AI / NLP
openai>=1.3.0
textblob>=0.17.1
nltk>=3.8.1
deep-translator>=1.11.4
newspaper3k>=0.2.8
feedparser>=6.0.10

# PDF 생성
reportlab>=4.0.0

# Excel 생성
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# 기타 유틸리티
lxml>=4.9.0
